+++ src/directives/index.ts
@@ -1,0 +1,2 @@
+export * from './withTime/withTime.directive';
+export * from './withToday/withToday.directive';